requirements.txt
fastapi>=0.98.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
python-jose>=3.3.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
PyJWT>=2.3.0,<3.0.0
pymysql>=1.0.0,<2.0.0
APScheduler>=3.9.0,<4.0.0